Split input into seperate files
